+++ R/Spectra.R
@@ -471,6 +471,9 @@ NULL
 #'   `rt[2]`. Returns the filtered `Spectra` (with spectra in their
 #'   original order).
 #'
+#' - `filterRanges`: allows filtering of the `Spectra` object
+#'    based on specified ranges for *any* values of `spectraVariables(object)`.
+#'
 #' - `reduceSpectra`: for groups of peaks within highly similar m/z values
 #'   within each spectrum (given `ppm` and `tolerance`), this function keeps
 #'   only the peak with the highest intensity removing all other peaks hence
@@ -956,6 +959,9 @@ NULL
 #' @param processingQueue For `Spectra`: optional `list` of
 #'     [ProcessingStep-class] objects.
 #'
+#' @param ranges for `filterRanges`: A `numeric` vector of paired values that
+#'     define the ranges to filter the spectra data.
+#'
 #' @param rt for `filterRt`: `numeric(2)` defining the retention time range to
 #'     be used to subset/filter `object`.
 #'
@@ -971,11 +977,15 @@ NULL
 #'     to import spectrum data from the provided files. See section *Creation
 #'     of objects, conversion and changing the backend* for more details.
 #'
-#' @param spectraVariables For `selectSpectraVariables`: `character` with the
+#' @param spectraVariables
+#' - For `selectSpectraVariables`: `character` with the
 #'     names of the spectra variables to which the backend should be subsetted.
-#'     For `addProcessing`: `character` with additional spectra variables that
+#' - For `addProcessing`: `character` with additional spectra variables that
 #'     should be passed along to the function defined with `FUN`. See function
 #'     description for details.
+#' - For `filterRanges`: A `character` or `integer` vector specifying the
+#'      column from `spectraVariables(object)` that correspond to the ranges
+#'      provided. The order must match the order of the parameter `ranges`.
 #'
 #' @param substDefinition For `deisotopeSpectra` and `filterPrecursorIsotopes`:
 #'     `matrix` or `data.frame` with definitions of isotopic substitutions.
@@ -1255,6 +1265,13 @@ NULL
 #' fft_spectrum_filtered
 #' length(mz(fft_spectrum_filtered)[[1]])
 #' plotSpectra(fft_spectrum_filtered, xlim = c(264.5, 265.5), ylim = c(0, 5e6))
+#'
+#' ## using filterRanges to filter spectra object based on variables available
+#' ## in `spectraData`.
+#' spectraVariables <- c("rtime", "precursorMz", "peaksCount")
+#' ranges <- c(30, 350, 200,500, 350, 600)
+#' filt_spectra <- filterRanges(data, spectraVariables = spectraVariables,
+#'                 ranges = ranges)
 #'
 #' ## ---- DATA MANIPULATIONS AND OTHER OPERATIONS ----
 #'
@@ -2395,6 +2412,42 @@ setMethod("reset", "Spectra", function(object, ...) {
     object@processing <- .logging(object@processing, "Reset object.")
     object
 })
+
+
+#' @rdname Spectra
+#' @importFrom MsCoreUtils between
+setMethod("filterRanges", "Spectra",
+          function(object, spectraVariables, ranges, ...){
+              if (is.logical(spectraVariables))
+                  spectraVariables <- which(spectraVariables)
+              if (length(spectraVariables) != length(ranges) / 2)
+                  stop("Length of 'spectraVariables' must be half the length ",
+                       "of 'ranges'")
+              if (is.character(spectraVariables))
+                  if(!all(spectraVariables %in% spectraVariables(object)))
+                  stop("'spectraVariables' need to correspond to colnames of",
+                  "the 'spectraData' of the object")
+              query <- spectraData(object)[, spectraVariables]
+              nc <- ncol(query)
+
+              within_ranges <- vapply(seq_len(nc), function(i) {
+                  pairs <-  c(ranges[2*i - 1], ranges[2*i])
+                  between(query[[i]], pairs)
+              }, logical(nrow(query)))
+
+              idc <- which(rowSums(within_ranges, na.rm = FALSE) == nc)
+              object@processing <- .logging(object@processing,
+                                            "Filter: select spectra with a ",
+                                            spectraVariables, " within: [",
+                                            ranges[seq(ranges)%% 2 == 0], ", ",
+                                            ranges[seq(ranges)%% 2 != 0], "]"
+                                            )
+              message("Started with ", length(object),
+                      " spectra, after filtering ", length(idc),
+                      " spectra are left")
+              object <- object[idc]
+          })
+
 
 #### ---------------------------------------------------------------------------
 ##