+++ tests/testthat/test_Spectra.R
@@ -1800,3 +1800,19 @@ test_that("entropy,Spectra works", {
   res <- entropy(sps, normalized = FALSE)
   expect_identical(res, vapply(df$intensity, MsCoreUtils::entropy, numeric(1)))
 })
+
+test_that("filterRanges works", {
+    filt_spectra <- filterRanges(sps_dia, spectraVariables = c("rtime",
+                                                               "precursorMz",
+                                                               "peaksCount"),
+                                 ranges = c(30, 350, 200,500, 350, 600))
+    logical_test <- spectraVariables(sps_dia) %in% c("rtime", "precursorMz",
+                                                     "peaksCount")
+    filt_spectra2 <- filterRanges(sps_dia, spectraVariables = logical_test,
+                                  ranges = c(30, 350, 200,500, 350, 600))
+    expect_true(length(sps_dia) > length(filt_spectra))
+    expect_error(filterRanges(sps_dia, spectraVariables = c("rtime",
+                                                             "precursorMz",
+                                                             "peaksCount")))
+    expect_error(filterRanges(sps_dia, ranges = c(30, 350, 200,500, 350, 600)))
+})